requirements.txt
flask
zxcvbn-python
gunicorn